requirements.txt
fastapi
uvicorn[standard]
streamlit
chromadb
langchain
langchain-huggingface
sentence-transformers
langchain_google_genai
sqlglot
sqlalchemy
python-dotenv
langgraph
tqdm
pydantic
requests
aiohttp
langchain-community
langchain-core